+++ clang/lib/CodeGen/CodeGenModule.cpp
@@ -3124,6 +3124,8 @@ void CodeGenModule::SetLLVMFunctionAttributesForDefinition(const Decl *D,
   if (std::optional<llvm::Attribute::AttrKind> Attr =
           StackProtectorAttribute(D)) {
     B.addAttribute(*Attr);
+    if (CodeGenOpts.StackProtectorLayoutOnly)
+      B.addAttribute("stack-protector-layout-only");
   }
 
   if (!D) {