+++ clang/docs/ReleaseNotes.md
@@ -239,6 +239,12 @@ features cannot lower the translation-unit ABI level;
 - New option `-fdefined-pointer-subtraction` added to preserve stable semantics
   when subtracting pointers to unrelated objects.
 
+- Added `-fstack-protector-layout-only`. When specified, changes to the layout
+  of a stack frame that occur when `-fstack-protector` is specified are applied
+  but a stack protector is not actually inserted. This allows targets that have
+  features that can act as pseudo-stack-protectors, like RISC-V's shadow stack,
+  to benefit from the protection added by reordering stack variables.
+
 - Added `--print-cxx-stdlib` and `--print-cxx-stdlib-include-dirs` to print
   the C++ standard library selected by the driver and the include directories
   added for it.